requirements.txt
Django==3.2.9
gunicorn==20.1.0
psycopg2-binary
rich
click
pyhcl
diagrams
